+++ rog-core/src/rogcore.rs
@@ -35,13 +35,14 @@ pub struct RogCore {
 impl RogCore {
     pub fn new(vendor: u16, product: u16, led_endpoint: u8) -> Result<RogCore, Box<dyn Error>> {
         let mut dev_handle = RogCore::get_device(vendor, product).map_err(|err| {
-            error!("Could not get device handle: {:?}", err);
+            error!("Could not get keyboard device handle: {:?}", err);
             err
         })?;
+        dev_handle.reset()?;
         dev_handle.set_active_configuration(0).unwrap_or(());
 
         let dev_config = dev_handle.device().config_descriptor(0).map_err(|err| {
-            error!("Could not get device config: {:?}", err);
+            error!("Could not get keyboard device config: {:?}", err);
             err
         })?;
         // Interface with outputs
@@ -69,7 +70,7 @@ impl RogCore {
                 err
             })?;
         dev_handle.claim_interface(interface).map_err(|err| {
-            error!("Could not claim device interface: {:?}", err);
+            error!("Could not claim keyboard device interface: {:?}", err);
             err
         })?;